+++ dashboard/views/account/view-buyer-order-invoice.php
@@ -226,7 +226,7 @@ if ($orderDetail['order_is_wallet_selected'] > 0) {
                     <table class="tbl-border" width="100%" border="0" cellpadding="10" cellspacing="0">
                         <thead>
                             <tr>
-                                <th width="<?php echo ($col6) ? '35%' : '55%'; ?>" style="padding:10px;text-align: left; border-bottom:1px solid #ddd;font-size:12px; background-color:#ddd;"><?php echo Labels::getLabel('LBL_Item', $siteLangId); ?></th>
+                                <th width="<?php echo ($col6) ? '35%' : '55%'; ?>" style="padding:10px;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?> border-bottom:1px solid #ddd;font-size:12px; background-color:#ddd;"><?php echo Labels::getLabel('LBL_Item', $siteLangId); ?></th>
                                 <th width="15%" style="padding:10px;text-align: center; border-bottom:1px solid #ddd;font-size:12px; background-color:#ddd;"><?php echo Labels::getLabel('LBL_Price', $siteLangId); ?></th>
                                 <th width="10%" style="padding:10px;text-align: center; border-bottom:1px solid #ddd;font-size:12px; background-color:#ddd;"><?php echo Labels::getLabel('LBL_Qty', $siteLangId); ?></th>
                                 <?php if ($col6) { ?>
@@ -243,39 +243,39 @@ if ($orderDetail['order_is_wallet_selected'] > 0) {
                         </thead>
                         <tbody>
                             <tr>
-                                <td style="padding:10px;text-align: left;"><?php echo $item; ?></td>
-                                <td style="padding:10px;text-align: left;"><?php echo CommonHelper::displayMoneyFormat($childOrder['op_unit_price'], true, false, true, false, true); ?></td>
-                                <td style="padding:10px;text-align: left;"><?php echo $childOrder['op_qty']; ?></td>
+                                <td style="padding:10px;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>"><?php echo $item; ?></td>
+                                <td style="padding:10px;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>"><?php echo CommonHelper::displayMoneyFormat($childOrder['op_unit_price'], true, false, true, false, true); ?></td>
+                                <td style="padding:10px;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>"><?php echo $childOrder['op_qty']; ?></td>
                                 <?php if ($col6) { ?>
-                                    <td style="padding:10px;text-align: left;"><?php echo CommonHelper::displayMoneyFormat($tax, true, false, true, false, true); ?></td>
+                                    <td style="padding:10px;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>"><?php echo CommonHelper::displayMoneyFormat($tax, true, false, true, false, true); ?></td>
                                 <?php } ?>
                                 <td style="padding:10px;text-align: center;font-size:12px;"><?php echo CommonHelper::displayMoneyFormat(CommonHelper::orderProductAmount($childOrder, 'CART_TOTAL'), true, false, true, false, true); ?></td>
                             </tr>
 
                             <tr>
-                                <td style="padding:10px; ;text-align: left;font-weight:700;background-color: #ddd;"><?php echo Labels::getLabel('Lbl_Summary', $siteLangId) ?> </td>
-                                <td style="padding:10px; ;text-align: left;background-color: #ddd;"><strong><?php echo CommonHelper::displayMoneyFormat($childOrder['op_unit_price'], true, false, true, false, true); ?></strong></td>
-                                <td style="padding:10px; ;text-align: left;background-color: #ddd;"><strong><?php echo $childOrder['op_qty']; ?></strong></td>
+                                <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>font-weight:700;background-color: #ddd;"><?php echo Labels::getLabel('Lbl_Summary', $siteLangId) ?> </td>
+                                <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>background-color: #ddd;"><strong><?php echo CommonHelper::displayMoneyFormat($childOrder['op_unit_price'], true, false, true, false, true); ?></strong></td>
+                                <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>background-color: #ddd;"><strong><?php echo $childOrder['op_qty']; ?></strong></td>
                                 <?php if ($col6) { ?>
-                                    <td style="padding:10px; ;text-align: left;background-color: #ddd;"><strong><?php echo CommonHelper::displayMoneyFormat($tax, true, false, true, false, true); ?></strong></td>
+                                    <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>background-color: #ddd;"><strong><?php echo CommonHelper::displayMoneyFormat($tax, true, false, true, false, true); ?></strong></td>
                                 <?php } ?>
                                 <td style="padding:10px; ;text-align: right;background-color: #ddd;"><strong><?php echo CommonHelper::displayMoneyFormat(CommonHelper::orderProductAmount($childOrder, 'CART_TOTAL'), true, false, true, false, true); ?></strong></td>
                             </tr>
                             <tr>
-                                <td style="padding:10px;font-size:15px;text-align: left;font-weight:700; vertical-align: top;" colspan="<?php echo ($col6) ? '2' : '1'; ?>" rowspan="10">
+                                <td style="padding:10px;font-size:15px;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>font-weight:700; vertical-align: top;" colspan="<?php echo ($col6) ? '2' : '1'; ?>" rowspan="10">
                                     <?php if (0 != $totalSavings) {
                                         $str = Labels::getLabel("LBL_You_have_saved_{totalsaving}_on_this_order", $siteLangId);
                                         $str = str_replace("{totalsaving}", CommonHelper::displayMoneyFormat(-$totalSavings, true, false, true, false, true), $str);
                                         echo $str;
                                     } ?>
                                 </td>
-                                <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('Lbl_Cart_Total', $siteLangId) ?></td>
+                                <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('Lbl_Cart_Total', $siteLangId) ?></td>
                                 <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo CommonHelper::displayMoneyFormat(CommonHelper::orderProductAmount($childOrder, 'cart_total'), true, false, true, false, true); ?></td>
                             </tr>
                             <?php
                             if ($volumeDiscount != 0) { ?>
                                 <tr>
-                                    <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_VOLUME_DISCOUNT', $siteLangId) ?></td>
+                                    <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_VOLUME_DISCOUNT', $siteLangId) ?></td>
                                     <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo '-' . CommonHelper::displayMoneyFormat($volumeDiscount, true, false, true, false, true); ?></td>
                                 </tr>
                                 <?php }
@@ -283,24 +283,24 @@ if ($orderDetail['order_is_wallet_selected'] > 0) {
                                 $rewardPointDiscount = CommonHelper::orderProductAmount($childOrder, 'REWARDPOINT');
                                 if ($rewardPointDiscount != 0) { ?>
                                     <tr>
-                                        <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Reward_Points', $siteLangId) ?></td>
+                                        <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Reward_Points', $siteLangId) ?></td>
                                         <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo CommonHelper::displayMoneyFormat($rewardPointDiscount, true, false, true, false, true); ?></td>
                                     </tr>
                                 <?php }
                                 if ($couponDiscount != 0) { ?>
                                     <tr>
-                                        <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Discount', $siteLangId) ?></td>
+                                        <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Discount', $siteLangId) ?></td>
                                         <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo '-' . CommonHelper::displayMoneyFormat($couponDiscount, true, false, true, false, true); ?></td>
                                     </tr>
                                 <?php }
                             }
                             if ($tax > 0) { ?>
                                 <tr>
-                                    <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_TAXABLE_AMOUNT', $siteLangId) ?></td>
+                                    <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_TAXABLE_AMOUNT', $siteLangId) ?></td>
                                     <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo CommonHelper::displayMoneyFormat($taxableAmount, true, false, true, false, true); ?></td>
                                 </tr>
                                 <tr>
-                                    <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Tax_Charges', $siteLangId) ?></td>
+                                    <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Tax_Charges', $siteLangId) ?></td>
                                     <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo CommonHelper::displayMoneyFormat($tax, true, false, true, false, true); ?></td>
                                 </tr>
                                 <?php }
@@ -308,26 +308,26 @@ if ($orderDetail['order_is_wallet_selected'] > 0) {
                                 $rewardPointDiscount = CommonHelper::orderProductAmount($childOrder, 'REWARDPOINT');
                                 if ($rewardPointDiscount != 0) { ?>
                                     <tr>
-                                        <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Reward_Points', $siteLangId) ?></td>
+                                        <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Reward_Points', $siteLangId) ?></td>
                                         <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo CommonHelper::displayMoneyFormat($rewardPointDiscount, true, false, true, false, true); ?></td>
                                     </tr>
                                 <?php }
                                 if ($couponDiscount != 0) { ?>
                                     <tr>
-                                        <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Discount', $siteLangId) ?></td>
+                                        <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Discount', $siteLangId) ?></td>
                                         <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo '-' . CommonHelper::displayMoneyFormat($couponDiscount, true, false, true, false, true); ?></td>
                                     </tr>
                                 <?php }
                             }
                             if ($childOrder['op_product_type'] != Product::PRODUCT_TYPE_DIGITAL && $childOrder['opshipping_fulfillment_type'] == Shipping::FULFILMENT_SHIP && 0 < CommonHelper::orderProductAmount($childOrder, 'shipping')) {  ?>
                                 <tr>
-                                    <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Delivery_Charges', $siteLangId) ?></td>
+                                    <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo Labels::getLabel('LBL_Delivery_Charges', $siteLangId) ?></td>
                                     <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><?php echo CommonHelper::displayMoneyFormat(CommonHelper::orderProductAmount($childOrder, 'shipping'), true, false, true, false, true); ?></td>
                                 </tr>
                             <?php }
                             if (array_key_exists('order_rounding_off', $orderDetail) && 0 != $orderDetail['order_rounding_off']) { ?>
                                 <tr>
-                                    <td style="padding:10px; ;text-align: left;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo (0 < $orderDetail['order_rounding_off']) ? Labels::getLabel('LBL_Rounding_Up', $siteLangId) : Labels::getLabel('LBL_Rounding_Down', $siteLangId); ?>
+                                    <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><?php echo (0 < $orderDetail['order_rounding_off']) ? Labels::getLabel('LBL_Rounding_Up', $siteLangId) : Labels::getLabel('LBL_Rounding_Down', $siteLangId); ?>
                                     </td>
                                     <td style="padding:10px;text-align: right;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1">
                                         <?php echo CommonHelper::displayMoneyFormat($orderDetail['order_rounding_off'], true, false, true, false, true); ?>
@@ -335,7 +335,7 @@ if ($orderDetail['order_is_wallet_selected'] > 0) {
                                 </tr>
                             <?php } ?>
                             <tr>
-                                <td style="padding:10px; ;text-align: left;font-weight:700;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><strong><?php echo Labels::getLabel('LBL_Grand_Total', $siteLangId) ?></strong> </td>
+                                <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?>font-weight:700;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="2"><strong><?php echo Labels::getLabel('LBL_Grand_Total', $siteLangId) ?></strong> </td>
                                 <td style="padding:10px; ;text-align: right;font-weight:700;border-top:1px solid #ddd;border-left:1px solid #ddd;" colspan="1"><strong><?php echo CommonHelper::displayMoneyFormat(CommonHelper::orderProductAmount($childOrder), true, false, true, false, true); ?></strong></td>
                             </tr>
                         </tbody>
@@ -366,7 +366,7 @@ if ($orderDetail['order_is_wallet_selected'] > 0) {
                                                 <?php }
                                                 } ?>
                                                 <tr>
-                                                    <td style="padding:10px; ;text-align: left; " colspan="2">*<?php echo Labels::getLabel('LBL_Appropriated_product-wise_and_Rate_applicable_thereunder', $siteLangId); ?></td>
+                                                    <td style="padding:10px; ;text-align: <?php echo CommonHelper::getLayoutDirection() == 'rtl' ? 'right' : 'left'; ?> " colspan="2">*<?php echo Labels::getLabel('LBL_Appropriated_product-wise_and_Rate_applicable_thereunder', $siteLangId); ?></td>
                                                 </tr>
                                             </tbody>
                                         </table>